Fixing type conversion issue.

+++ src/carl/core/MultivariatePolynomial.h
@@ -323,7 +323,7 @@ public:
      * @return An approximation of the complexity of this polynomial.
      */
     std::size_t complexity() const {
-		return std::accumulate(begin(), end(), 0,
+		return std::accumulate(begin(), end(), static_cast<std::size_t>(0),
 			[](std::size_t cur, const auto& t){ return cur + t.complexity(); }
 		);
     }
